+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for a second criteria to match the hover image beyond the label option.
+
+### Fixed
+- Main product image was still appearing when a hover image with transparency was being displayed.
 
 ## [2.67.0] - 2020-12-17
 ### Deprecated